changed spot data type to decimal for lat/long

+++ backend/db/migrations/20230805205136-create-spot.js
@@ -41,11 +41,11 @@ module.exports = {
           allowNull: false,
         },
         lat: {
-          type: Sequelize.FLOAT,
+          type: Sequelize.DECIMAL,
           allowNull: false,
         },
         lng: {
-          type: Sequelize.FLOAT,
+          type: Sequelize.DECIMAL,
           allowNull: false,
         },
         name: {
